[SYSTEMDS-3940] Improve Scuro Node Executor
In this patch the node executor for Representation DAGs is improved and made mode efficient. Additionally, the code was cleaned up.

+++ src/main/python/tests/scuro/data_generator.py
@@ -222,7 +222,9 @@ class ModalityRandomDataGenerator:
         data = [
             [
                 random.random()
-                for _ in range(random.randint(max_audio_length * 0.9, max_audio_length))
+                for _ in range(
+                    random.randint(int(max_audio_length * 0.9), max_audio_length)
+                )
             ]
             for _ in range(num_instances)
         ]
